requirements.txt
requests
tabulate
python-bidi
